break out PodMatchingContext, more unit tests
